+++ src/main/java/com/salesforce/dynamodbv2/mt/mappers/sharedtable/impl/MtAmazonDynamoDbBySharedTable.java
@@ -12,6 +12,7 @@ import static com.google.common.base.Preconditions.checkArgument;
 import static java.util.stream.Collectors.toList;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
+import com.amazonaws.services.dynamodbv2.model.AttributeDefinition;
 import com.amazonaws.services.dynamodbv2.model.AttributeValue;
 import com.amazonaws.services.dynamodbv2.model.BatchGetItemRequest;
 import com.amazonaws.services.dynamodbv2.model.BatchGetItemResult;
@@ -32,6 +33,7 @@ import com.amazonaws.services.dynamodbv2.model.PutItemRequest;
 import com.amazonaws.services.dynamodbv2.model.PutItemResult;
 import com.amazonaws.services.dynamodbv2.model.QueryRequest;
 import com.amazonaws.services.dynamodbv2.model.QueryResult;
+import com.amazonaws.services.dynamodbv2.model.ScalarAttributeType;
 import com.amazonaws.services.dynamodbv2.model.ScanRequest;
 import com.amazonaws.services.dynamodbv2.model.ScanResult;
 import com.amazonaws.services.dynamodbv2.model.StreamSpecification;
@@ -47,7 +49,6 @@ import com.salesforce.dynamodbv2.mt.context.MtAmazonDynamoDbContextProvider;
 import com.salesforce.dynamodbv2.mt.mappers.MtAmazonDynamoDbBase;
 import com.salesforce.dynamodbv2.mt.mappers.metadata.DynamoTableDescriptionImpl;
 import com.salesforce.dynamodbv2.mt.mappers.metadata.PrimaryKey;
-import com.salesforce.dynamodbv2.mt.mappers.sharedtable.impl.FieldPrefixFunction.FieldValue;
 import com.salesforce.dynamodbv2.mt.repo.MtTableDescriptionRepo;
 import com.salesforce.dynamodbv2.mt.util.StreamArn;
 import java.time.Clock;
@@ -135,7 +136,7 @@ public class MtAmazonDynamoDbBySharedTable extends MtAmazonDynamoDbBase {
         return mtTables.containsKey(tableName);
     }
 
-    Function<Map<String, AttributeValue>, FieldValue> getFieldValueFunction(String sharedTableName) {
+    Function<Map<String, AttributeValue>, FieldValue<?>> getFieldValueFunction(String sharedTableName) {
         CreateTableRequest table = mtTables.get(sharedTableName);
         checkArgument(table != null);
         // TODO consider representing physical tables as DynamoTableDescription
@@ -143,9 +144,19 @@ public class MtAmazonDynamoDbBySharedTable extends MtAmazonDynamoDbBase {
                 .filter(elem -> HASH.toString().equals(elem.getKeyType()))
                 .map(KeySchemaElement::getAttributeName)
                 .findFirst().orElseThrow(IllegalStateException::new);
-        FieldPrefixFunction fpf = new FieldPrefixFunction('.');
-        // TODO support non-string physical table hash key
-        return key -> fpf.reverse(key.get(hashKeyName).getS());
+        ScalarAttributeType hashKeyType = table.getAttributeDefinitions().stream()
+            .filter(attr -> hashKeyName.equals(attr.getAttributeName()))
+            .map(AttributeDefinition::getAttributeType)
+            .map(ScalarAttributeType::valueOf)
+            .findFirst().orElseThrow(IllegalStateException::new);
+        switch (hashKeyType) {
+            case S:
+                return key -> StringFieldPrefixFunction.INSTANCE.reverse(key.get(hashKeyName).getS());
+            case B:
+                return key -> BinaryFieldPrefixFunction.INSTANCE.reverse(key.get(hashKeyName).getB());
+            default:
+                throw new IllegalStateException("Unsupported physical table hash key type " + hashKeyType);
+        }
     }
 
     @Override
@@ -158,6 +169,10 @@ public class MtAmazonDynamoDbBySharedTable extends MtAmazonDynamoDbBase {
      */
     @Override
     public BatchGetItemResult batchGetItem(BatchGetItemRequest unqualifiedBatchGetItemRequest) {
+        // validate
+        unqualifiedBatchGetItemRequest.getRequestItems().values()
+            .forEach(MtAmazonDynamoDbBySharedTable::validateGetItemKeysAndAttribute);
+
         // clone request and clear items
         Map<String, KeysAndAttributes> unqualifiedKeysByTable = unqualifiedBatchGetItemRequest.getRequestItems();
         BatchGetItemRequest qualifiedBatchGetItemRequest = unqualifiedBatchGetItemRequest.clone();
@@ -194,9 +209,35 @@ public class MtAmazonDynamoDbBySharedTable extends MtAmazonDynamoDbBase {
                 tableMapping.getVirtualTable().getTableName(),
                 qualifiedItems.stream().map(keysAndAttributes ->
                     tableMapping.getItemMapper().reverse(keysAndAttributes)).collect(Collectors.toList()));
+            // map unprocessedKeys
+            if (!qualifiedBatchGetItemResult.getUnprocessedKeys().isEmpty()) {
+                unqualifiedBatchGetItemResult.clearUnprocessedKeysEntries();
+                qualifiedBatchGetItemResult.getUnprocessedKeys()
+                    .forEach((qualifiedTableNameUk, qualifiedUkKeysAndAttributes) -> {
+                        TableMapping tableMappingUk = tableMappingByPhysicalTableName.get(qualifiedTableNameUk);
+                        unqualifiedBatchGetItemResult.addUnprocessedKeysEntry(
+                            tableMappingUk.getVirtualTable().getTableName(),
+                            new KeysAndAttributes()
+                                .withKeys(qualifiedUkKeysAndAttributes.getKeys().stream()
+                                .map(keysAndAttributes ->
+                                    tableMapping.getKeyMapper().reverse(keysAndAttributes))
+                                    .collect(Collectors.toList())));
+                    });
+            }
         });
 
         return unqualifiedBatchGetItemResult;
+    }
+
+    private static void validateGetItemKeysAndAttribute(KeysAndAttributes keysAndAttributes) {
+        checkArgument(keysAndAttributes.getConsistentRead() == null,
+            "setting consistentRead is not supported on BatchGetItemRequest calls");
+        checkArgument(keysAndAttributes.getAttributesToGet() == null,
+            "attributesToGet are not supported on BatchGetItemRequest calls");
+        checkArgument(keysAndAttributes.getProjectionExpression() == null,
+            "projectionExpression is not supported on BatchGetItemRequest calls");
+        checkArgument(keysAndAttributes.getExpressionAttributeNames() == null,
+            "expressionAttributeNames are not supported on BatchGetItemRequest calls");
     }
 
     /**
@@ -267,6 +308,16 @@ public class MtAmazonDynamoDbBySharedTable extends MtAmazonDynamoDbBase {
      */
     @Override
     public GetItemResult getItem(GetItemRequest getItemRequest) {
+        // validate
+        checkArgument(getItemRequest.getConsistentRead() == null,
+            "setting consistentRead is not supported on GetItemRequest calls");
+        checkArgument(getItemRequest.getAttributesToGet() == null,
+            "attributesToGet are not supported on GetItemRequest calls");
+        checkArgument(getItemRequest.getProjectionExpression() == null,
+            "projectionExpression is not supported on GetItemRequest calls");
+        checkArgument(getItemRequest.getExpressionAttributeNames() == null,
+            "expressionAttributeNames are not supported on GetItemRequest calls");
+
         // map table name
         getItemRequest = getItemRequest.clone();
         TableMapping tableMapping = getTableMapping(getItemRequest.getTableName());