+++ src/test/java/com/salesforce/dynamodbv2/mt/mappers/MtAmazonDynamoDbChainTest.java
@@ -62,6 +62,7 @@ import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -82,6 +83,7 @@ class MtAmazonDynamoDbChainTest {
     private static final int TIMEOUT_SECONDS = 600;
     private static final boolean IS_LOCAL_DYNAMO = true;
 
+    @Disabled
     @Test
     void test() throws Exception {
         // create context