+++ src/test/java/com/salesforce/dynamodbv2/mt/mappers/sharedtable/impl/FieldPrefixFunctionTest.java
@@ -7,14 +7,12 @@
 
 package com.salesforce.dynamodbv2.mt.mappers.sharedtable.impl;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.salesforce.dynamodbv2.mt.context.MtAmazonDynamoDbContextProvider;
-import com.salesforce.dynamodbv2.mt.mappers.sharedtable.impl.FieldPrefixFunction.FieldValue;
-
+import java.nio.ByteBuffer;
 import java.util.Arrays;
 import java.util.stream.Stream;
-
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
@@ -25,35 +23,41 @@ import org.junit.jupiter.params.provider.MethodSource;
  */
 class FieldPrefixFunctionTest {
 
-    private static final FieldPrefixFunction SUT = new FieldPrefixFunction('.');
+    private static Object[] binaryData(String context, String table, String value) {
+        return new Object[] {
+            BinaryFieldPrefixFunction.INSTANCE, context, table, UTF_8.encode(value), ByteBuffer
+            .allocate(context.length() + table.length() + value.length() + 2)
+            .put(UTF_8.encode(context)).put((byte) 0x00)
+            .put(UTF_8.encode(table)).put((byte) 0x00)
+            .put(UTF_8.encode(value))
+            .flip()
+        };
+    }
 
     static Stream<Object[]> data() {
         return Arrays.stream(new Object[][] {
-            { "ctx", "table", "value", "ctx.table.value" },
-            { "ctx2", "com.salesforce.zero.someObject", "value2", "ctx2.com\\.salesforce\\.zero\\.someObject.value2" }
+            { StringFieldPrefixFunction.INSTANCE, "ctx", "table", "value",
+                "ctx/table/value" },
+            { StringFieldPrefixFunction.INSTANCE, "ctx2", "com.salesforce.zero.someObject", "value2",
+                "ctx2/com.salesforce.zero.someObject/value2" },
+            { StringFieldPrefixFunction.INSTANCE, "ctx_3", "My-Object", "prefix/suffix",
+                "ctx_3/My-Object/prefix/suffix" },
+            binaryData("ctx", "table", "value"),
+            binaryData("ctx", "table", "val\u0000ue"),
         });
     }
 
     @ParameterizedTest
     @MethodSource("data")
-    void applyAndReverse(String context, String tableIndex, String unqualifiedValue, String qualifiedValue) {
-        FieldValue expected = new FieldValue(context, tableIndex, qualifiedValue, unqualifiedValue);
+    <V> void applyAndReverse(FieldPrefixFunction<V> sut, String context, String tableIndex, V unqualifiedValue,
+                             V qualifiedValue) {
+        FieldValue<V> expected = new FieldValue<>(context, tableIndex, unqualifiedValue);
 
-        FieldValue applied = SUT.apply(new MtAmazonDynamoDbContextProvider() {
-            @Override
-            public String getContext() {
-                return context;
-            }
+        V actual = sut.apply(expected);
 
-            @Override
-            public void withContext(String org, Runnable runnable) {
+        assertEquals(qualifiedValue, actual);
 
-            }
-        }, tableIndex, unqualifiedValue);
-
-        assertEquals(expected, applied);
-
-        assertEquals(expected, SUT.reverse(applied.getQualifiedValue()));
+        assertEquals(expected, sut.reverse(actual));
     }
 
 }